Factor out benchmark creation code into different module.
